move wagmi init to index page
